Fixed buffer errors in tests.
There were a number of usages of `sprintf' that potentially allowed
buffer overflows to occur. While this may not be a big issue,
converting them to `snprintf' calls allows them to generally be safer.

Also, on Arch Linux (as of 2018-01-03) the test wouldn't compile
because the compiler determined that there were some instances where
the buffer was too small. So changed the buffer size from 1024 to 2048
in those instances.