Fix slot scanning and acquiring password.
The slot scanning array was initialized to 0, and then processes were
being forked off to populate it. However, because it is a fork, the
information does not return to the parent process and the parent version
of the array remains at 0 for all elements. 0 is interpreted as a
successful result of unlocking the slot with a given passphrase, so slot
0 was always selected as an up-to-date slot.

By removing the fork behavior, this begins to work again. Defensively,
setting the initial version of the array to all -1 is also an
improvement for future bugs of this nature. It is pretty slow to check
all the slots, and the forking functionality could be later replaced
with threading support.

After fixing this, the slot was being set with the user's name as a
passphrase rather than the actual password they had. This is because we
were grabbing PAM_USER for both username and password. By grabbing
PAM_AUTHTOK for the password, we are able to correctly set the right
passphrase into the slot.